phase-3/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlmodel==0.0.14
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Phase-3 Chat API dependencies
openai>=1.0.0
asyncpg>=0.29.0
httpx>=0.25.0

# Phase-3 MCP Task Server dependencies
mcp[cli]
